requirements.txt
blinker==1.7.0
boto3==1.34.59
botocore==1.34.59
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.3
ffmpeg-python==0.2.0
Flask==3.0.2
future==1.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==1.26.4
packaging==23.2
pillow==10.2.0
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
urllib3==2.0.7
Werkzeug==3.0.1
